Update: re-add our redis caching logic to some controllers as we commented it out for testing/local dev purposes. update the notification service app.js cors so we can actually communicatewith the FE and vice versa. Add: redis caching logic to notifications/get/set and clearing.

+++ story-service/src/utils/cleanRedisCache.utils.js
@@ -1,5 +1,9 @@
-export async function clearRedisStoriesCache(req) {
-  const keys = await req.redisClient.keys(`stories`);
+export async function clearRedisStoriesCache(req, userId) {
+  if (!userId) {
+    return;
+  }
+
+  const keys = await req.redisClient.keys(`stories:${userId}`);
 
   if (Array.isArray(keys) && keys.length > 0) {
     await req.redisClient.unlink(keys);
